Move LtiMessageStore to LTI::MessageStore
